Feat: BaseballGameService 서비스 추가
